# api-docs.yaml
openapi: 3.0.1
info:
  title: iBoard 社區公佈欄(後端Spring boot)
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /board/{boardId}:
    get:
      tags:
      - 公佈欄 CRUD
      summary: 取得公佈欄資料
      description: 透過board_id取得公佈欄資料
      operationId: getBoardById
      parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
    put:
      tags:
      - 公佈欄 CRUD
      summary: 修改公佈欄訊息
      description: 透過 boardId 查詢訊息後修改訊息
      operationId: updateBoard
      parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardRequest'
        required: true
      responses:
        "404":
          description: 無此 公佈欄訊息 ID，無法執行修改功能
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
    delete:
      tags:
      - 公佈欄 CRUD
      summary: 刪除公佈欄訊息
      description: 透過 boardId 刪除已存在訊息
      operationId: deleteBoard
      parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
  /admins/{adminId}:
    get:
      tags:
      - 管理員 CRUD
      summary: 透過adminId查詢管理員資料
      description: 透過adminId查詢管理員資料
      operationId: getAdminById
      parameters:
      - name: adminId
        in: path
        description: 管理員 ID 編號
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Admin'
        "404":
          description: '無此管理員 ID '
    put:
      tags:
      - 管理員 CRUD
      summary: 修改管理員資料
      description: 透過adminId先查詢是否存在，如存在該筆數據，執行修改；最後再使用adminId重新查詢該筆數據後回傳
      operationId: updateAdmin
      parameters:
      - name: adminId
        in: path
        description: 管理員 ID 編號
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        description: 請求參數列表
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRequest'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Admin'
        "404":
          description: 無此管理員 ID
    delete:
      tags:
      - 管理員 CRUD
      summary: 刪除管理員資料
      description: 透過adminId刪除管理員資料
      operationId: deleteAdmin
      parameters:
      - name: adminId
        in: path
        description: 管理員 ID 編號
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
...
  /board:
    get:
      tags:
      - 公佈欄 CRUD
      summary: 取得所有公佈欄資訊
      description: 取得所有公佈欄資訊
      operationId: getBoards
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
    post:
      tags:
      - 公佈欄 CRUD
      summary: 新增公佈欄訊息
      description: 新增公佈欄訊息
      operationId: createBoard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardRequest'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
  /admins:
    get:
      tags:
      - 管理員 CRUD
      summary: 取得所有管理員
      description: 取得所有管理員資料
      operationId: getAdmins
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Admin'
    post:
      tags:
      - 管理員 CRUD
      summary: 新增管理員資料
      description: 新增管理員資料
      operationId: createAdmin
      requestBody:
        description: 請求參數列表
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRequest'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Admin'
...
components:
  schemas:
...
    BoardRequest:
      required:
      - boardCategory
      - boardDescription
      - boardTitle
      type: object
      properties:
        boardCategory:
          type: string
          description: 訊息分類
          enum:
          - IMPORTANT
          - ADVERTISE
          - ACTIVITY
          - ANNOUNCEMENT
        boardTitle:
          type: string
          description: 公告title
        boardDescription:
          type: string
          description: 公告內容
    Board:
      type: object
      properties:
        boardId:
          type: integer
          description: 公佈欄 ID
          format: int32
        boardCategory:
          type: string
          description: 公佈欄 分類
          enum:
          - IMPORTANT
          - ADVERTISE
          - ACTIVITY
          - ANNOUNCEMENT
        boardTitle:
          type: string
          description: '公佈欄 標題 '
        boardDescription:
          type: string
          description: 公佈欄 內容文字
        createdDate:
          type: string
          description: 公佈欄 被創建時間
          format: date-time
        lastModifiedDate:
          type: string
          description: 公佈欄 最後被修改時間
          format: date-time
    AdminRequest:
      required:
      - adminAccount
      - adminAddress
      - adminEmail
      - adminName
      - adminPassword
      - adminTel
      type: object
      properties:
        adminAccount:
          type: string
          description: 管理員帳號
        adminPassword:
          type: string
          description: 管理員密碼
        adminName:
          type: string
          description: 管理員姓名
        adminTel:
          type: string
          description: 管理員電話
        adminEmail:
          type: string
          description: 管理員信箱
        adminAddress:
          type: string
          description: 管理員地址
    Admin:
      type: object
      properties:
        adminId:
          type: integer
          description: 管理員 ID
          format: int32
        adminAccount:
          type: string
          description: 管理員帳號
        adminPassword:
          type: string
          description: 管理員密碼
        adminName:
          type: string
          description: 管理員姓名
        adminTel:
          type: string
          description: 管理員電話
        adminEmail:
          type: string
          description: 管理員 email
        adminAddress:
          type: string
          description: 管理員地址
        createdDate:
          type: string
          description: 管理員資料被創建時間
          format: date-time
        lastModifiedDate:
          type: string
          description: 管理員資料最後修改時間
          format: date-time
...